moved packages and better enabled name
