+++ src/main/java/at/hannibal2/skyhanni/config/features/combat/FlareConfig.java
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.config.features.misc;
+package at.hannibal2.skyhanni.config.features.combat;
 
 import at.hannibal2.skyhanni.config.FeatureToggle;
 import at.hannibal2.skyhanni.config.core.config.Position;
@@ -12,10 +12,10 @@ import io.github.notenoughupdates.moulconfig.annotations.ConfigOption;
 public class FlareConfig {
 
     @Expose
-    @ConfigOption(name = "Enable Overlay", desc = "Show current active flares.")
+    @ConfigOption(name = "Enable", desc = "Show current active flares.")
     @ConfigEditorBoolean
     @FeatureToggle
-    public boolean overlayEnabled = false;
+    public boolean enabled = false;
 
     @Expose
     @ConfigOption(name = "Alert Type", desc = "What type of alert should be sent when a flare is about to expire.")
